+++ plugwise/__init__.py
@@ -1,6 +1,6 @@
 """Plugwise module."""
 
-__version__ = "0.21.3"
+__version__ = "0.22.0a"
 
 from plugwise.smile import Smile
 from plugwise.stick import Stick